test: dependencies test
